refactor: Seperated MetadataSelect into own component.
